I am going to hell

+++ src/main/java/com/Harbinger/Spore/Sentities/InfectedVendicator.java
@@ -96,7 +96,7 @@ public class InfectedVendicator extends EvolvedInfected {
             AABB aabb = this.getBoundingBox().inflate(0.2);
             for(BlockPos blockpos : BlockPos.betweenClosed(Mth.floor(aabb.minX), Mth.floor(aabb.minY), Mth.floor(aabb.minZ), Mth.floor(aabb.maxX), Mth.floor(aabb.maxY), Mth.floor(aabb.maxZ))) {
                 BlockState blockstate = this.level.getBlockState(blockpos);
-                if (blockstate.getMaterial() == Material.WOOD && this.isAggressive()) {
+                if (blockstate.getMaterial() == Material.WOOD && this.isAggressive() && blockstate.getDestroySpeed(level ,blockpos) < 2) {
                     flag = this.level.destroyBlock(blockpos, true, this) || flag;
                 }
             }
